+++ src/Mscc.GenerativeAI/Enums/FinishReason.cs
@@ -2,9 +2,12 @@
 using System.Text.Json.Serialization;
 #endif
 
-
 namespace Mscc.GenerativeAI
 {
+    /// <summary>
+    /// The reason why the model stopped generating tokens.
+    /// If empty, the model has not stopped generating the tokens.
+    /// </summary>
     [JsonConverter(typeof(JsonStringEnumConverter<FinishReason>))]
 
     public enum FinishReason
@@ -12,7 +15,7 @@ namespace Mscc.GenerativeAI
         /// <summary>
         /// Unspecified means the finish reason is unspecified.
         /// </summary>
-        Unspecified = 0,
+        FinishReasonUnspecified = 0,
         /// <summary>
         /// Stop means natural stop point of the model or provided stop sequence.
         /// </summary>
@@ -35,6 +38,21 @@ namespace Mscc.GenerativeAI
         /// <summary>
         /// Other means all other reasons that stopped the token generation
         /// </summary>
-        Other = 5
+        Other = 5,
+        /// <summary>
+        /// The token generation was stopped as the response was flagged for the
+        /// terms which are included from the terminology blocklist.
+        /// </summary>
+        Blocklist = 6,
+        /// <summary>
+        /// The token generation was stopped as the response was flagged for
+        /// the prohibited contents.
+        /// </summary>
+        ProhibitedContent = 7,
+        /// <summary>
+        /// The token generation was stopped as the response was flagged for
+        /// Sensitive Personally Identifiable Information (SPII) contents.
+        /// </summary>
+        Spii = 8
     }
 }