+++ src/Mscc.GenerativeAI/Enums/HarmCategory.cs
@@ -1,6 +1,5 @@
-﻿
+﻿#if NET472_OR_GREATER || NETSTANDARD2_0
 using System;
-#if NET472_OR_GREATER || NETSTANDARD2_0
 using System.Text.Json.Serialization;
 #endif