+++ src/Mscc.GenerativeAI/CHANGELOG.md
@@ -13,6 +13,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Feature suggestion: Add logs with LogLevel using the Standard logging in .NET ([#6](https://github.com/mscraftsman/generative-ai/issues/6))
 - implement Automatic Function Call (AFC)
 - implement Server-Sent Events (SSE)
+
+### Changed
+### Fixed
+
+## 1.1.4
+
+### Added
+
 - new values in enum FinishReason
 - new enum HarmBlockMethod
 
@@ -21,8 +29,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - improve enums (ref: Google.Cloud.AIPlatform.V1)
 - improve response in SSE format
 - update samples to latest NuGet package
-
-### Fixed
 
 ## 1.1.3